Authorization - added authorization to controllers

+++ VolleyLeague.API/Controllers/LeagueController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using VolleyLeague.Services.Interfaces;
 using VolleyLeague.Shared.Dtos.Teams;
 
@@ -18,6 +19,7 @@ namespace VolleyLeague.API.Controllers
             _leagueService = leagueService;
         }
 
+        [AllowAnonymous]
         [HttpGet("GetAllLeagues")]
         public async Task<IActionResult> GetAllLeagues()
         {
@@ -25,30 +27,45 @@ namespace VolleyLeague.API.Controllers
             return Ok(result);
         }
 
+        [Authorize(Roles = "Admin")]
         [HttpPost("CreateLeague")]
         public async Task<IActionResult> CreateLeague([FromBody] LeagueDto league)
         {
+            string? idUser = User.Identity?.Name;
+            if (string.IsNullOrWhiteSpace(idUser))
+            {
+                return Unauthorized();
+            }
+
             await _leagueService.CreateLeague(league);
             return Ok();
         }
 
+        [Authorize(Roles = "Admin")]
         [HttpDelete("DeleteLeague")]
         public async Task<IActionResult> DeleteLeague(int id)
         {
+            string? idUser = User.Identity?.Name;
+            if (string.IsNullOrWhiteSpace(idUser))
+            {
+                return Unauthorized();
+            }
+
             var result = await _leagueService.DeleteLeague(id);
             if (result)
                 return Ok();
             else return BadRequest();
         }
 
+        [Authorize(Roles = "Admin")]
         [HttpPut("UpdateLeague")]
         public async Task<IActionResult> UpdateLeague([FromBody] LeagueDto league)
         {
-            //string? id = User.Identity?.Name;
-            //if (string.IsNullOrWhiteSpace(id))
-            //{
-            //    return Unauthorized();
-            //}
+            string? id = User.Identity?.Name;
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                return Unauthorized();
+            }
             await _leagueService.UpdateLeague(league);
 
             return Ok();
